Changed download location of NEON data; fixed bug in sixs validation when more than two built exes
