+++ isofit/data/cli/sixs.py
@@ -2,6 +2,7 @@
 Downloads 6S from https://github.com/isofit/6S
 """
 
+import logging
 import os
 import platform
 import shutil
@@ -22,6 +23,56 @@ from isofit.data.download import (
 ESSENTIAL = True
 CMD = "sixs"
 MINGW = "https://github.com/brechtsanders/winlibs_mingw/releases/download/15.2.0posix-13.0.0-msvcrt-r2/winlibs-i686-posix-dwarf-gcc-15.2.0-mingw-w64msvcrt-13.0.0-r2.zip"
+
+Logger = logging.getLogger(__name__)
+
+
+def get_exe(path: str = None, version: bool = False) -> str:
+    """
+    Retrieves the 6S executable from a given path
+
+    Parameters
+    ----------
+    path : str, default=None
+        6S directory path. If None, defaults to the ini sixs path
+    version : bool, default=False
+        Returns the 6S version instead
+
+    Returns
+    -------
+    pathlib.Path | str
+        Either the 6S executable as a pathlib object or the string 6S version
+    """
+    if path is None:
+        path = env.sixs
+
+    path = Path(path)
+
+    exes = path.glob("sixsV*")
+    exes = [exe for exe in exes if "lutaero" not in exe.name]
+    names = [exe.name for exe in exes]
+
+    if not exes:
+        raise FileNotFoundError(f"Could not find a 6S executable under path: {path}")
+
+    if len(exes) > 1:
+        Logger.warning(
+            f"More than one 6S executable was found. Defaulting to the first one: {names}"
+        )
+
+    if version:
+        # Try using the version.txt file, created by the isofit downloader
+        if (txt := path / "version.txt").exists():
+            with txt.open("r") as f:
+                vers = f.read()
+        # Fallback to using the executable name
+        else:
+            _, vers = names[0].split("V")
+            vers = f"v{vers}".lower()
+
+        return vers
+
+    return exes[0]
 
 
 def precheck():
@@ -224,7 +275,9 @@ def validate(path=None, checkForUpdate=True, debug=print, error=print, **_):
         error("[x] 6S path does not exist")
         return False
 
-    if len(list(path.glob("sixsV2*"))) != 2:
+    try:
+        exe = get_exe(path)
+    except FileNotFoundError:
         error(
             "[x] 6S is missing the built 'sixsV2.*', this is likely caused by make failing"
         )
